fix: change communication_pause and removed rounding operator (//) for communication_pause and timeout (timeout was rounded down to 0, if baudrate was higher than 20000)

+++ src/tmc_driver/com/_tmc_com.py
@@ -56,7 +56,7 @@ class TmcCom:
         """constructor."""
         self._tmc_logger: TmcLogger
         self._driver_address: int = 0
-        self.communication_pause: int = 0
+        self.communication_pause: float = 0
         self._get_register_callback = None
 
     @abstractmethod
